Big number support

+++ src/Service/EthereumServiceInterface.php
@@ -12,14 +12,14 @@ interface EthereumServiceInterface
 
     public function getEtherBalance(string $address): float;
 
-    public function getTokenBalance(string $contract, string $address): int;
+    public function getTokenBalance(string $contract, string $address): float;
 
     public function sendEther(string $from, string $to, float $value): string;
 
     public function getTransactionReceipt(string $txHash): ?array;
 
-    public function transferToken(string $tokenContract, string $from, string $to, int $value): string;
+    public function transferToken(string $tokenContract, string $from, string $to, float $value): string;
 
-    public function approveTransfer(string $tokenContract, string $from, string $spender, int $value): string;
+    public function approveTransfer(string $tokenContract, string $from, string $spender, float $value): string;
 
 }
